Delete function working done

+++ task/views.py
@@ -53,8 +53,9 @@ def task_detail(request, task_id):
 @login_required
 def task_delete(request, task_id):
     task = get_object_or_404(Task, id=task_id, user=request.user)
-    task.delete()
-    return render(request,"task_list.html")
+    if request.method == "POST":
+        task.delete()
+    return redirect("task_list")
 @login_required
 def task_mark_completed(request,task_id):
     task = get_object_or_404(Task, id=task_id, user=request.user)
